+++ src/components/Notifications/index.js
@@ -5,6 +5,7 @@ import { format, utcToZonedTime } from 'date-fns-tz';
 import { GlobalsContext } from 'Shared/context';
 import Notification from './Notification';
 import { addNotification } from './actions';
+import { getDesktopOrMobile } from '../../helpers';
 
 import style from './style.scss';
 
@@ -30,12 +31,12 @@ const Notifications = ({
   useEffect(() => {
     let timeout;
 
-    if (sessionDurationInSec > 1800) { // 30 minutes * 60 seconds
+    if (sessionDurationInSec > 600) { // 10 minutes * 60 seconds
       // date (yyyy-mm) is stored in UTC so it needs to be checked in UTC as well
       const date = format(utcToZonedTime(new Date(), 'UTC'), 'yyyy-MM', { timeZone: 'UTC' });
       const existingNotification = notifications.some(({ id }) => id === date);
 
-      if (!rate[date] && !existingNotification) {
+      if (!rate[`${date}-${getDesktopOrMobile()}`] && !existingNotification) {
         timeout = setTimeout(() => addNotificationAction({ id: date, type: 'rate' }), 1000);
       }
     }